refactor: ref directive
