+++ src/index.ts
@@ -10,4 +10,5 @@ export * from './presets/classic/components'
 export * from './presets/context-menu/components'
 export * from './presets/minimap/components'
 export * from './presets/reroute/components'
+export * from './ref'
 export * from './core'
