+++ src/module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { NodeComponent, RefDirective } from './presets/classic/components/node/node.component';
+import { NodeComponent } from './presets/classic/components/node/node.component';
 import { ConnectionComponent } from './presets/classic/components/connection/connection.component';
 import { SocketComponent } from './presets/classic/components/socket/socket.component';
 import { ConnectionWrapperComponent } from './presets/classic/components/connection/connection-wrapper.component';
 import { ControlComponent } from './presets/classic/components/control/control.component';
+import { RefDirective } from './ref';
 
 @NgModule({
   declarations: [
@@ -35,4 +36,4 @@ import { ControlComponent } from './presets/classic/components/control/control.c
     ControlComponent
   ]
 })
-export class ReteModule {}
+export class ReteModule { }
